Gooey init commit
Changed  argument parser to gooey.  Preparing to allow for both CLI and GUI arg parsers.
