+++ deeplc/gui.py
@@ -1,0 +1,7 @@
+"""Graphical user interface."""
+
+from gooey import GooeyParser
+
+from deeplc.__main__ import main
+
+
